+++ src/iris-and-literals.ts
@@ -1,0 +1,34 @@
+const semtools = require('semantic-toolkit');
+
+import { getPrefixes } from './prefixes';
+import * as Helpers from './helpers';
+
+// TODO: Literals
+export async function getIRIsAndLiterals(ontology): Promise<Object> {
+  const triples = Helpers.getTriples(ontology);
+  const prefixes = await getPrefixes(ontology);
+  const prefixMap = Helpers.invertObject(prefixes);
+
+  const graph = triples.reduce((memo, triple) => {
+    const { subject, predicate, object } = triple;
+    return [ ...memo, subject, predicate, object];
+  }, []).reduce((memo, maybeIri) => {
+    if (!semtools.isIri(maybeIri)) return memo;
+
+    const prefix = semtools.getNamespace(maybeIri);
+    const localName = semtools.getLocalName(maybeIri);
+    const prefixLocalName = prefixMap[prefix];
+    // console.log('Prefix and localname:', prefix, localName);
+
+    if (!(prefixLocalName in memo)) memo[prefixLocalName] = {};
+    memo[prefixLocalName][localName] = maybeIri;
+
+    return memo;
+  }, {});
+
+  return {
+    exports: [ 'graph' ],
+    iris: graph,
+    literals: {}
+  }
+}
